backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.35.0
httpx>=0.25.0
beautifulsoup4>=4.12.2
tldextract>=5.1.0
requests>=2.31.0
urllib3>=2.0.7
sqlalchemy>=2.0.23
alembic>=1.13.1
email-validator>=2.1.0